fix: add missing team_invitation_team session key
The ProcessPendingTeamInvitation middleware expects 'team_invitation_team' to be stored in the session, but the TeamInvitationController wasn't setting it.

This caused the success message to show 'null' instead of the team name when a user registered/logged in after clicking an invitation link.

Fixed by adding 'team_invitation_team' => \->team->name to the session data.